A package to help you decide what talks to attend at JSM

We wrangled the JSM program and decades of citation and coauthorship data from Semantic Scholar and Arxiv. We have also included functionality for exporting your schedule as an ical that can be loaded into Google Calendar or similar. The result is a package that streamlines the process of finding talks that you may want to attend.

See also

Author

Maintainer: Jacob Bien jbien@usc.edu

Other contributors:

Yibin Xiong [contributor]